# docs/swagger.yaml
basePath: /
definitions:
  core.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  coverage.CoverageListResponse:
    properties:
      coverages:
        items:
          $ref: '#/definitions/coverage.CoverageResponse'
        type: array
      repo:
        $ref: '#/definitions/core.Repository'
    type: object
...
  coverage.CoverageUploadRequest:
    properties:
      entries:
        items:
          $ref: '#/definitions/coverage.CoverageEntryUploadRequest'
        type: array
      revision:
        type: string
      time:
        type: string
    type: object
...
  server.PasswordLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  server.User:
    properties:
      avatar_url:
        type: string
      id:
        type: integer
      user_type:
        type: string
      username:
        type: string
    type: object
...
  server.createCoverageTrackerRequest:
    properties:
      name:
        type: string
      repo_id:
        type: integer
      visibility:
        type: string
    type: object
...
  tracker.TrackerModel:
    properties:
      body:
        type: string
      chart_config:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      last_updated_at:
        type: string
      name:
        type: string
      owner_id:
        type: integer
      type:
        type: string
      visibility:
        type: string
    type: object
...
info:
  contact: {}
  description: Mora is a code coverage tracker API.
  title: Mora API
  version: "1.0"
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with a username and password, creating a session
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.PasswordLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Log in with username and password
      tags:
      - auth
...
  /api/coverages:
    post:
      consumes:
      - application/json
      description: Create a coverage-type tracker linked to a repository. The repository
        must not already have a coverage tracker.
      parameters:
      - description: Coverage tracker information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.createCoverageTrackerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tracker.TrackerModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Create a coverage tracker
      tags:
      - coverage
  /api/coverages/{trackerId}:
    get:
      description: Return coverage history for a tracker (public list endpoint)
      parameters:
      - description: Tracker ID
        in: path
        name: trackerId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/coverage.CoverageListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: List coverages for a tracker
      tags:
      - coverage
    post:
      consumes:
      - application/json
      description: Upload coverage entries for a repository linked to a tracker
      parameters:
      - description: Tracker ID
        in: path
        name: trackerId
        required: true
        type: integer
      - description: Coverage data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/coverage.CoverageUploadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/coverage.Coverage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Upload coverage data
      tags:
      - coverage
...
  /api/coverages/{trackerId}/preview:
    get:
      description: Return coverage timeline data as virtual series for a tracker
      parameters:
      - description: Tracker ID
        in: path
        name: trackerId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tracker.PreviewResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Preview coverage data
      tags:
      - coverage
  /api/me:
    get:
      description: Return the current session user, or 204 if not logged in
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.User'
        "204":
          description: No Content
      summary: Get current user
      tags:
      - server
  /api/providers:
    get:
      description: Return configured SCM repository managers and their login status
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.RepositoryManagerResponse'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: List SCM providers
      tags:
      - server
  /api/repos:
    get:
      description: List repositories the current session can access
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.Repository'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: List repositories
      tags:
      - server
...
  /api/trackers:
    get:
      description: Return trackers owned, edited, or liked by the current user. With
        include=preview, each tracker also embeds its preview data (all series plus
        their latest values) fetched in a few batched queries.
      parameters:
      - description: Set to \
        in: query
        name: include
        type: string
      - description: Search query (partial match on tracker name)
        in: query
        name: q
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tracker.ListTrackersResponse'
      summary: List trackers for current user
      tags:
      - server
    post:
      consumes:
      - application/json
      description: Add a new tracker. The name must be unique. Creator becomes owner.
      parameters:
      - description: Tracker information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tracker.CreateTrackerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tracker.TrackerModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Create a tracker
      tags:
      - tracker
  /api/trackers/{trackerId}:
    delete:
      description: Delete the specified tracker. Child series and values are also
        cascade-deleted. Owner only.
      parameters:
      - description: Tracker ID
        in: path
        name: trackerId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Delete a tracker
      tags:
      - tracker
    get:
      description: Return a single tracker by ID
      parameters:
      - description: Tracker ID
        in: path
        name: trackerId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tracker.TrackerResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Get a tracker
      tags:
      - tracker
    patch:
      consumes:
      - application/json
      description: Update tracker fields (e.g. visibility). Requires owner permission.
      parameters:
      - description: Tracker ID
        in: path
        name: trackerId
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tracker.PatchTrackerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tracker.TrackerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Update a tracker
      tags:
      - tracker
...
  /api/user/me/api-keys/{keyId}:
    delete:
      description: Revoke an API key by ID
      parameters:
      - description: API key ID
        in: path
        name: keyId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Revoke an API key
      tags:
      - api-key
  /api/users/{userName}:
    get:
      description: Return the public user profile for the given username
      parameters:
      - description: Username
        in: path
        name: userName
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Get a user by username
      tags:
      - server
...
swagger: "2.0"